requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
google-generativeai==0.3.1
python-dotenv==1.0.0
Werkzeug==2.3.7
Pillow==10.0.1
uuid==1.30 